fix: Correct backward-compatibile method aliases (#35)
The simple method aliasing in #34 didn't actually work. Calling one of those methods results in a TypeError saying `this.endpoint` is not a function, which makes sense in retrospect, since `endpoint` isn't bound to the constructed object.

To actually get the claimed backward compatibility (and avoid further problems with `this`), the alias methods are now real methods that forward their arguments to the namespaced ones.