fix: Unpack axios errors for better debugging (#37)
When using a client with an invalid Stytch env (for example, local testing without the server running), the error messages from endpoint methods were difficult to work with (the dreaded "[object Object]"). To make these debuggable, we now unpack the axios error into useful parts as a dedicated `RequestError` type.