dev: Bring back unsendable-request test (#52)
This test was removed without deep investigation during #51 because it was blocking the release.  My current hypothesis is that we were relying on internal axios error message strings, which broke in the upgrade from v0.21.1 to v0.21.3.

To make this test more reliable and more representative of a realistic error, force the BigInt serialization problem. As a bonus, if we can ever start relying on BigInts to serialize, this test will fail, which will prompt us to consider using them again.