dev: Run linter and formatter in CI (#67)
I noticed that we weren't enforcing these in CI while merging the recent ESLint and Prettier updates. With these changes, they'll now run as part of a new Actions workflow.

As part of this, I ran Prettier on everything in `lib` to make the check pass.

Relevant commands for dev:
* `npm run lint`: Run ESLint (already existed)
* `npm run format`: Format all code with Prettier in write mode
* `npm run check-format`: Run Prettier in check mode

This PR contains an intermediate commit with a CI test failure to validate the check output.