ci: Ensure the CI project always has a JWT signing key (#161)
The sandbox tests started failing in CI recently. The API was unable to sign the sandbox-session JWT for the CI project because the original signing key expired.

Normally, the key would automatically be rotated when it was necessary to sign the JWT for a real session. However, the CI project only ever issues sandbox requests and never starts a real session. So this uses the workaround of fetching the JWKS to trigger signing key generation outside the sandbox.

This is a _very_ specific edge case that would only be encountered by a very new or very old Stytch project whose first interaction with sessions is to try and validate the sandbox token. You should never have to use this anywhere in your real projects, Test or Live.